+++ CHANGES.md
@@ -19,6 +19,7 @@ UI:
 -   Allow a blacklist of strings to be specified for automatic keyword generation
 -   Make the global notification banner configurable
 -   Removed all references to the DTA Design System react components, so that all styles come through our SCSS compilation, which should make SCSS smaller and more consistent.
+-   Swap the order of custodian & team dropdown on add dataset page
 
 Storage:
 
@@ -26,6 +27,7 @@ Storage:
 -   Add a DELETE endpoint
 -   Improves error handling (returns 404 from GET if the file doesn't exist)
 -   Add apidocs
+-   Add endpoint to create a bucket
 -   Restrict file upload to admins only
 
 Gateway:
@@ -35,6 +37,10 @@ Gateway:
 Gateway:
 
 -   Add /data to ckan URL, remove the `came_from` param
+
+Authorization:
+
+-   Made integration tests for authorisation run automatically as part of CI.
 
 Others: